+++ server/app.js
@@ -9,6 +9,10 @@ var csrf = require('csurf');
 var passport = require('passport');
 var logger = require('morgan');
 
+var ensureLogIn = require('connect-ensure-login').ensureLoggedIn;
+
+var ensureLoggedIn = ensureLogIn();
+
 // pass the session to the connect sqlite3 module
 // allowing it to inherit from session.Store
 var SQLiteStore = require('connect-sqlite3')(session);
@@ -49,8 +53,41 @@ app.use(function(req, res, next) {
   next();
 });
 
+//const baseUrl = process.env.BASE_URL;
 app.use('/', indexRouter);
 app.use('/', authRouter);
+
+// redirect any other route back to home route /
+// app.use((req,res,next)=>{
+// 	res.redirect('/');
+// });
+
+// Serve static files from the public folder with the base URL
+//app.use(baseUrl, express.static('public'));
+app.use('/home',express.static('../dist'));
+app.use('/public',express.static('../dist/public'));
+app.use('/public', express.static('public'));
+
+
+app.get('/client', function (req, res, next){
+	// get env variables for header
+	var CLOWDER_REMOTE_HOSTNAME = process.env.CLOWDER_REMOTE_HOSTNAME;
+	var APIKEY = process.env.APIKEY;
+	var options = {
+		headers:{
+			'hostname': CLOWDER_REMOTE_HOSTNAME,
+			'apikey': APIKEY
+		}
+	}
+	res.json(options); // Use this in src/utils/common in getHeader() method.
+});
+
+// app.get('/home', ensureLoggedIn, function (req, res, next){
+// 	// load build directory only if logged in
+// 	app.use('/home/',express.static('../dist'));
+// 	app.use('/public',express.static('../dist/public'));
+// 	res.sendFile(path.join(__dirname, '../dist', 'index.html'));
+// });
 
 // catch 404 and forward to error handler
 app.use(function(req, res, next) {