+++ src/actions/client.js
@@ -3,6 +3,7 @@
 import {createEmptyDatasetRequest, getFileInDataset, uploadFileToDatasetRequest} from "../utils/dataset";
 import {checkExtractionStatusLoop, submitForExtraction} from "../utils/file";
 import config from "../app.config";
+import {getClientInfo} from "../utils/common";
 import {SET_EXTRACTION_STATUS, setExtractionStatus} from "./file";
 import {ADD_FILE_TO_DATASET, addFileToDataset, CREATE_DATASETS, createDataset} from "./dataset";
 
@@ -11,22 +12,23 @@ import {ADD_FILE_TO_DATASET, addFileToDataset, CREATE_DATASETS, createDataset} f
 export function createUploadExtract(file) {
 	return async function createUploadExtractThunk(dispatch) {
 		// this function creates an empty dataset. uploads the file to the dataset and submits for extraction
+		const clientInfo = await getClientInfo();
 		// Clowder API call to create empty dataset
 		const file_name = file.name.replace(/\.[^/.]+$/, ""); // get filename without extension as dataset name
 		const file_description = file.type;
-		const dataset_json = await createEmptyDatasetRequest(file_name, file_description); // returns the dataset ID {id:xxx}
-		if (dataset_json !== undefined) {
+		const dataset_json = await createEmptyDatasetRequest(file_name, file_description, clientInfo); // returns the dataset ID {id:xxx}
+		if (dataset_json !== undefined && dataset_json !== null) {
 			dispatch(createDataset(CREATE_DATASETS, dataset_json));
 			// upload input file to dataset
-			let file_json = await uploadFileToDatasetRequest(dataset_json.id, file); // return file ID. {id:xxx} OR {ids:[{id:xxx}, {id:xxx}]}
+			let file_json = await uploadFileToDatasetRequest(dataset_json.id, file, clientInfo); // return file ID. {id:xxx} OR {ids:[{id:xxx}, {id:xxx}]}
 			if (file_json !== undefined){
 				file_json["filename"] = file.name;
 				// submit uploaded file for extraction
 				if (file.type == "text/plain"){
-					const rct_extraction_submission = await submitForExtraction(file_json.id, config.rct_extractor);
+					const rct_extraction_submission = await submitForExtraction(file_json.id, config.rct_extractor, clientInfo);
 					if (rct_extraction_submission) {
 						// check every 5s for extraction status
-						//const rct_extraction_status = await checkExtractionStatusLoop(file_json.id, config.rct_extractor, 5000);
+						//const rct_extraction_status = await checkExtractionStatusLoop(file_json.id, config.rct_extractor, 5000, clientInfo);
 						const rct_extraction_status = true
 						if (rct_extraction_status === true){
 							console.log("RCT extraction status true");
@@ -45,19 +47,20 @@ export function createUploadExtract(file) {
 
 				}
 				else if (file.type == "application/pdf") {
-					const pdf_extraction_submission = await submitForExtraction(file_json.id, config.pdf_extractor);
+					const pdf_extraction_submission = await submitForExtraction(file_json.id, config.pdf_extractor, clientInfo);
 					if (pdf_extraction_submission) {
-						//const pdf_extraction_status = await checkExtractionStatusLoop(file_json.id, config.pdf_extractor, 5000);
+						//const pdf_extraction_status = await checkExtractionStatusLoop(file_json.id, config.pdf_extractor, 5000, clientInfo);
 						const pdf_extraction_status = true;
 						if (pdf_extraction_status === true){
 							console.log("pdf extraction done");
 							const text_file_name = file_name + '.txt';
-							const extracted_txt_file = await getFileInDataset(dataset_json.id, "text/file", text_file_name);
+							const extracted_txt_file = await getFileInDataset(dataset_json.id, "text/file", text_file_name, clientInfo);
 							if (extracted_txt_file !== null && typeof extracted_txt_file.id === "string") {
-								const rct_extraction_submission = await submitForExtraction(extracted_txt_file.id, config.rct_extractor);
+								const rct_extraction_submission = await submitForExtraction(extracted_txt_file.id, config.rct_extractor, clientInfo);
 								if (rct_extraction_submission === true){
 									// check every 5s for extraction status
-									const rct_extraction_status = await checkExtractionStatusLoop(extracted_txt_file.id, config.rct_extractor, 5000);
+									//const rct_extraction_status = await checkExtractionStatusLoop(extracted_txt_file.id, config.rct_extractor, 5000, clientInfo);
+									const rct_extraction_status = true;
 									if (rct_extraction_status === true){
 										console.log("RCT extraction status true");
 										dispatch(setExtractionStatus(SET_EXTRACTION_STATUS, true));