+++ src/components/childComponents/FilePreview.js
@@ -10,6 +10,9 @@ import Video from "../previewers/Video";
 import Thumbnail from "../previewers/Thumbnail";
 import {getPreviewResources} from "../../utils/file";
 import PreviewDrawerLeft from "./PreviewDrawerLeft";
+import Intro from "./Intro";
+import CreateAndUpload from "./CreateAndUpload";
+import {getClientInfo} from "../../utils/common";
 
 
 export default function FilePreview() {
@@ -25,7 +28,8 @@ export default function FilePreview() {
 			const previewsTemp = [];
 			filePreviews[0].map(async (preview) => {
 				// get all preview resources
-				const preview_config = await getPreviewResources(preview);
+				const clientInfo = await getClientInfo()
+				const preview_config = await getPreviewResources(preview, clientInfo);
 				previewsTemp.push(preview_config);
 				setPreviews(previewsTemp); // set previews
 			});