+++ src/routes.tsx
@@ -9,7 +9,7 @@ export const AppRoutes = (): JSX.Element => {
 	return (
 		<BrowserRouter>
 			<Routes>
-				<Route path="/" element={<Dashboard/>}/>
+				<Route path="/home/" element={<Dashboard/>}/>
 				<Route path="/preview" element={<Preview/>} />
 				<Route path="*"
 					   element={