+++ src/utils/common.js
@@ -1,21 +1,36 @@
+import axios from "axios"
 import config from "../app.config";
 
+// get client endpoint
+const getClient = {method:'GET', url:"/client"};
+
+export function getClientInfo(){
+	return axios.request(getClient).then(function (response) {
+		return response.data.headers
+	});
+}
+
+// TODO remove this method
+// get hostname
+export function getHostname(){
+	return axios.request(getClient).then(function (response) {
+		return response.data.headers.hostname
+	});
+}
 
 // construct header
-export function getHeader() {
+export function getHeader(clientInfo) {
 	const headers = new Headers({
-		"X-API-Key": config.apikey
+		"X-API-Key": clientInfo.apikey
 	});
-
 	return headers;
-
 	// const headers = new Headers({
 	// 	"Authorization": cookies.get("Authorization"),
 	// });
 }
 
-export async function downloadResource(url) {
-	let authHeader = getHeader();
+export async function downloadResource(url, clientInfo) {
+	let authHeader = getHeader(clientInfo);
 	let response = await fetch(url, {
 		method: "GET",
 		mode: "cors",