+++ CHANGELOG.md
@@ -15,6 +15,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Client can handle pdf file uploads and triggers appropriate extractors. [#31](https://github.com/clowder-framework/CONSORT-frontend/issues/31)
 - Client checks the extraction status and repeat extraction submission if needed. [#34](https://github.com/clowder-framework/CONSORT-frontend/issues/34)
 - Render metadata in Html file preview. [#39](https://github.com/clowder-framework/CONSORT-frontend/issues/39)
+- Express server calling react client. [#49](https://github.com/clowder-framework/CONSORT-frontend/issues/49)
+- Dockerize frontend. [#51](https://github.com/clowder-framework/CONSORT-frontend/issues/51)
 
 ### Changed
 - Status message display in overlay. [#26](https://github.com/clowder-framework/CONSORT-frontend/issues/26)